pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multiagent_hackathon"
version = "0.1.0"
description = "A multi-agent hackathon project"
requires-python = ">=3.11"
dependencies=[
    "ipython",
    "langchain",
    "python-dotenv",
    "autogen-agentchat==0.4.4",
    "autogen-ext[openai, azure]",
    "flaml[automl]",
    "mammoth",
    "markdownify",
    "python-pptx",
    "puremagic",
    "pathvalidate",
    "pdfminer",
    "ruff",
    "yfinance",
    "matplotlib",
    "langchain-azure-dynamic-sessions==0.2.0",
    "semantic-kernel==1.24.0",
    "fastapi[standard]",
]

[tool.ruff]
line-length = 79  
